feat(help): In editor help viewer.
* Add project templates for Skyrim LE

* Add generate skyrim LE project files menu options

* Add copyAndFillTemplate()
Simple template mechanism

* Oops, wrong filename

* no debug support for skyrimLE

* Return promise, don't await

* Add Skyrim LE support and improve code a bit

* Add global constants file

This is for any sort of miscellaneous global constants.

* add WelcomeHandler

At some point we might also want to show a "what's new" when
an upgrade happens. This is where the code for handling that
will go. For now it doesn't really do much.

* add ShowWelcomeCommand (markdown)

Add a command to show a markdown preview of our
welcome and getting started document.

* Add menu item to show getting started help again

* First version of the Getting Started document

* Add debugging, some formatting

* Cleanup, reenable "only display once"

* Fix nasty typo in build comand

* Only show view/title menu items for enabled games

* Fix nasty bug: projectFolder set too early

* Add global constants file

This is for any sort of miscellaneous global constants.

* add WelcomeHandler

At some point we might also want to show a "what's new" when
an upgrade happens. This is where the code for handling that
will go. For now it doesn't really do much.

* add ShowWelcomeCommand (markdown)

Add a command to show a markdown preview of our
welcome and getting started document.

* Add menu item to show getting started help again

* First version of the Getting Started document

* Add debugging, some formatting

* Cleanup, reenable "only display once"

* Fix nasty typo in build comand

* Remove duplicate menu item