Replace Paths.ts with PathResolver (ioc service)
This is a new service that attempts to handle all possible
issues with regard to path resolution so we don't have so
many dependencies on config, context, etc, all over the place.
It simplifies the code elsewhere by eliminating some of the
IOC dependencies that get included purely for path resolution.