Fix for incorrect files being opened is vscode when running server in mono
When running the papyrus-lang dotnet server in mono, it appears there is an incompatibility where it will mangle the file:/ descriptor. It will add the mountpoint to the beginning of the uri, which obviously will break the editor as it will try to open a non-existent file. Add a guard against that code block when running in mono, to instead use the working file:///, which appears to work in mono and vscode running under linux.